1 down, and fixed page shift
Finished my part of the tags page, as well as fixed a problem with the page shifting. It was due to the scroll bar appearing on the right which moved everything to the left. Added a few lines to the css so that this bar always appears and everything stays the same.